requirements.txt
# Core dependencies
langchain==0.0.267
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
langchain-pinecone>=0.1.0

# Document processing
pymupdf==1.23.0
pdfplumber==0.10.2
pypdf==3.15.1

# Vector database
pinecone-client==2.2.4
sentence-transformers==2.2.2

# Web interface
streamlit==1.28.0

# Development
python-dotenv==1.0.0
pytest==7.4.0